+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- [PR-56](https://github.com/itk-dev/aapodwalk/pull/56)
+  - Rename component `Map.jsx` -> `MapComponent.jsx` to "[not shadow the global "Map" property](https://eslint.org/docs/latest/rules/no-shadow)"
+  - Create a mapper to map lats/longs to fit the outer bounds react-leaflet uses
+  - Create a `MapMarker.jsx`
 - [PR-55](https://github.com/itk-dev/aapodwalk/pull/55)
   - Redesign pins on map
 - [PR-54](https://github.com/itk-dev/aapodwalk/pull/54)